Move password control to its own folder
